Relocate migrations and improve error handle
