+++ packages/dappmanager/src/index.ts
@@ -5,9 +5,6 @@ import initializeDb from "./initializeDb";
 import { createGlobalEnvsEnvFile } from "./modules/globalEnvs";
 import { generateKeyPair } from "./utils/publickeyEncryption";
 import { copyHostScripts } from "./modules/hostScripts";
-import { switchEthClientIfOpenethereumOrGethLight } from "./modules/ethClient";
-import { runLegacyActions } from "./modules/legacy";
-import { migrateUserActionLogs } from "./logUserAction";
 import { postRestartPatch } from "./modules/installer/restartPatch";
 import { startDaemons } from "./daemons";
 import { SshManager } from "./modules/sshManager";
@@ -24,10 +21,9 @@ import {
 } from "./utils/getVersionData";
 import { shellHost } from "./utils/shell";
 import { startDappmanager } from "./startDappmanager";
-import { addAliasToRunningContainersMigration } from "./modules/https-portal";
 import { copyHostServices } from "./modules/hostServices/copyHostServices";
 import { startAvahiDaemon } from "./daemons/avahi";
-import { setDefaultStakerConfig } from "./modules/stakerConfig/setDefaultStakerConfig";
+import { executeMigrations } from "./modules/migrations";
 
 const controller = new AbortController();
 
@@ -49,11 +45,8 @@ const server = startDappmanager({
   sshManager
 });
 
-// Deprecate openethereum. MUST be executed before EthClientInstaller daemon to avoid have the error:
-// `Error on eth client installer daemon Error: No client data for target: openethereum`
-switchEthClientIfOpenethereumOrGethLight().catch(e =>
-  logs.error("Error deprecating ethereum client: ", e)
-);
+// Execute migrations
+executeMigrations().catch(e => logs.error("Error on executeMigrations", e));
 
 // Initialize DB
 initializeDb()
@@ -100,30 +93,6 @@ calls
 const versionData = getVersionData();
 if (versionData.ok) logs.info("Version info", versionData.data);
 else logs.error(`Error getting version data: ${versionData.message}`);
-
-/**
- * [LEGACY] The previous method of injecting ENVs to a DNP was via .env files
- * This function will read the contents of .env files and add them in the
- * compose itself in the `environment` field in array format.
- *
- * [LEGACY] The DB is split into two where the old db becomes a cache only
- * and the new one is for permanent required data. Some key-values will be
- * moved from the old db to the cache db.
- */
-
-migrateUserActionLogs().catch(e =>
-  logs.error("Error migrating userActionLogs", e)
-);
-
-runLegacyActions().catch(e => logs.error("Error running legacy actions", e));
-
-addAliasToRunningContainersMigration().catch(e =>
-  logs.error("Error adding alias to running containers", e)
-);
-
-setDefaultStakerConfig().catch(e =>
-  logs.error("Error setting default staker config", e)
-);
 
 postRestartPatch().catch(e => logs.error("Error on postRestartPatch", e));
