+++ packages/dappmanager/src/modules/ethClient/index.ts
@@ -1,4 +1,3 @@
 export * from "./changeEthMultiClient";
 export * from "./ethersProvider";
-export * from "./switchEthClientIfOpenethereumOrGethLight";
 export * from "./localFallbackVersions";
