+++ packages/dappmanager/test/integration/migrationCompose/migration.test.int.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import fs from "fs";
 import { PackageContainer } from "../../../src/common";
-import { migrateCoreNetworkAndAliasInCompose } from "../../../src/modules/https-portal/migration";
+import { migrateCoreNetworkAndAliasInCompose } from "../../../src/modules/migrations/addAliasToRunningContainers";
 import params from "../../../src/params";
 import { mockContainer, shellSafe } from "../../testUtils";
